back-end/requirements.txt
Django==4.2
django-versatileimagefield==2.2
django-ckeditor==6.5.1
django-colorfield==0.4.1
djangorestframework==3.14.0
psycopg2-binary==2.8.6
django-filter==23.1
